+++ src/test/java/lumi/insert/app/controller/transactionpayment/TransactionPaymentControllerGetTest.java
@@ -8,6 +8,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import java.math.BigDecimal;
 import java.util.List;
  
 import org.junit.jupiter.api.DisplayName;
@@ -108,7 +109,7 @@ public class TransactionPaymentControllerGetTest extends BaseTransactionPaymentC
     @DisplayName("Should return slice of transactionPayment DTO when transaction found/valid")
     void searchTransactionPaymentsFilterAPI_validRequest_shouldReturnSliceEntityDTO() throws Exception{
         Slice<TransactionPaymentResponse> slice = new SliceImpl<TransactionPaymentResponse>(List.of(transactionPaymentResponse));
-        when(transactionPaymentService.getTransactionPaymentsByRequests(TransactionPaymentGetByFilter.builder().transactionId(transactionPaymentResponse.transactionId()).minTotalPayment(5000L).build())).thenReturn(slice);
+        when(transactionPaymentService.getTransactionPaymentsByRequests(TransactionPaymentGetByFilter.builder().transactionId(transactionPaymentResponse.transactionId()).minTotalPayment(BigDecimal.valueOf(5000L)).build())).thenReturn(slice);
 
         mockMvc.perform(
             get("/api/transactions/" + transactionPaymentResponse.transactionId() + "/payments/filter?minTotalPayment=5000")
@@ -125,7 +126,7 @@ public class TransactionPaymentControllerGetTest extends BaseTransactionPaymentC
     @DisplayName("Should return error missmatch when request param type missmatch")
     void searchTransactionPaymentsFilterAPI_missmatchParam_shouldReturnSliceEntityDTO() throws Exception{
         Slice<TransactionPaymentResponse> slice = new SliceImpl<TransactionPaymentResponse>(List.of(transactionPaymentResponse));
-        when(transactionPaymentService.getTransactionPaymentsByRequests(TransactionPaymentGetByFilter.builder().transactionId(transactionPaymentResponse.transactionId()).minTotalPayment(5000L).build())).thenReturn(slice);
+        when(transactionPaymentService.getTransactionPaymentsByRequests(TransactionPaymentGetByFilter.builder().transactionId(transactionPaymentResponse.transactionId()).minTotalPayment(BigDecimal.valueOf(5000L)).build())).thenReturn(slice);
 
         mockMvc.perform(
             get("/api/transactions/" + transactionPaymentResponse.transactionId() + "/payments/filter?minTotalPayment=true")
@@ -140,7 +141,7 @@ public class TransactionPaymentControllerGetTest extends BaseTransactionPaymentC
     @DisplayName("Should return error badrequest  when param filter totalpayment is lower than 0")
     void searchTransactionPaymentsFilterAPI_minTotalPaymentMinus_shouldReturnSliceEntityDTO() throws Exception{
         Slice<TransactionPaymentResponse> slice = new SliceImpl<TransactionPaymentResponse>(List.of(transactionPaymentResponse));
-        when(transactionPaymentService.getTransactionPaymentsByRequests(TransactionPaymentGetByFilter.builder().transactionId(transactionPaymentResponse.transactionId()).minTotalPayment(5000L).build())).thenReturn(slice);
+        when(transactionPaymentService.getTransactionPaymentsByRequests(TransactionPaymentGetByFilter.builder().transactionId(transactionPaymentResponse.transactionId()).minTotalPayment(BigDecimal.valueOf(5000L)).build())).thenReturn(slice);
 
         mockMvc.perform(
             get("/api/transactions/" + transactionPaymentResponse.transactionId() + "/payments/filter?minTotalPayment=-5000")
@@ -155,7 +156,7 @@ public class TransactionPaymentControllerGetTest extends BaseTransactionPaymentC
     @DisplayName("Should return error badrequest  when param filter totalpayment is lower than 0")
     void searchTransactionPaymentsFilterAPI_wrongSortByValue_shouldReturnSliceEntityDTO() throws Exception{
         Slice<TransactionPaymentResponse> slice = new SliceImpl<TransactionPaymentResponse>(List.of(transactionPaymentResponse));
-        when(transactionPaymentService.getTransactionPaymentsByRequests(TransactionPaymentGetByFilter.builder().transactionId(transactionPaymentResponse.transactionId()).minTotalPayment(5000L).build())).thenReturn(slice);
+        when(transactionPaymentService.getTransactionPaymentsByRequests(TransactionPaymentGetByFilter.builder().transactionId(transactionPaymentResponse.transactionId()).minTotalPayment(BigDecimal.valueOf(5000L)).build())).thenReturn(slice);
 
         mockMvc.perform(
             get("/api/transactions/" + transactionPaymentResponse.transactionId() + "/payments/filter?minTotalPayment=5000&sortBy=cc")
